src/bankd/bankd_main: fix error on debian 13 i386
Fix for:

  bankd_main.c:824:46: error: passing argument 5 of 'worker->ops->transceive' from incompatible pointer type [-Wincompatible-pointer-types]
    824 |                                      rx_buf, &rx_buf_len);
        |                                              ^~~~~~~~~~~
        |                                              |
        |                                              DWORD * {aka long unsigned int *}
  bankd_main.c:824:46: note: expected 'size_t *' {aka 'unsigned int *'} but argument is of type 'DWORD *' {aka 'long unsigned int *'}

Related: OS#6828

+++ src/bankd/bankd_main.c
@@ -792,7 +792,7 @@ static int worker_handle_tpduModemToCard(struct bankd_worker *worker, const Rspr
 {
 	const struct TpduModemToCard *mdm2sim = &pdu->msg.choice.tpduModemToCard;
 	uint8_t rx_buf[1024];
-	DWORD rx_buf_len = sizeof(rx_buf);
+	size_t rx_buf_len = sizeof(rx_buf);
 	RsproPDU_t *pdu_resp;
 	struct client_slot clslot;
 	struct bank_slot bslot;
